refactor: move DEFAULT_CUSTOM_THEME to its own constants file
DEFAULT_CUSTOM_THEME (a prayer-timings custom theme default) lived in
slide-design.ts, which is about Ayat/Hadith slides. Move it to
constants/custom-theme.ts; consumers import via the @/constants barrel and
are unaffected.
